+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.4.5 (unreleased)
 
 - Improved error message for non-string values
+- Fixed error with migrating Action Text
 
 ## 0.4.4 (2020-06-23)